Revert previous commit, fix properly
Previous commit set the made initial bet flag in next player,
which gets called in places where the flag shouldn't be set, for
example, when the players make forced small blind/big blind.
These of course are not their initial bets, and they can still
bet if they choose.

Now it is set in promptbet.